.screenrc: altscreen=on (alternate buffer)
xterm-style emulators support the notion of an alternate buffer with
a separate scroll-back history than the "normal" buffer. By default
GNU screen has the alternate buffer disabled, but we want it enabled
so that alternate buffer switching works as expected (when using an
xterm-supporting terminal emulator) for applications that try to use
that it, e.g. vim, etc.