.bashrc: Cleanup $PATH handling, remove unused cases
Build discrete "start" and "end" lists, and prepend/append those lists
to $PATH/$MANPATH once at the end. This makes it easier to control the
discrete order of paths in the prefix vs suffix sections.  For example,
I always want $HOME/bin to be at the very-start of $PATH so that I can
put an "override" script in $HOME/bin and that can always win over
"normal" binaries in other later directories in $PATH.